+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes will be documented in this file.
 
 ### Changed
 
+* Reduced logging (#72)
+* Logon timeout (#70)
 * Subscribe to reference data changes (FIX) (#65)
 * Added HTTP `request_id` (#55)
 * Base64 clean-up to support web-safe (#43)